requirements.txt
flask>=3.0.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langdetect>=1.0.9
requests>=2.31.0
